fix(field): handle missing interval view gracefully
The field should still work, even if the django_interval.urls are not
included in the projects urlpatterns. Therefore we put the `reverse`
method in a try/except block and only set the `data-intervaluri` if the
endpoint exists.
If the `data-intervaluri` is not set, the triggered javascript simply
does nothing.

Closes: #11