fix(views): catch exception and pass error message on to template
If the parsing of the datestring fails we catch the exception instead of
showing a error trace to the user. We then write the exception message
in the context, this way we can show it to the user.
This also updates the template to include the error message if there is
one.

Closes: #53