docs(fields): expand FuzzyDateRegexField documentation
Improve docstring for `FuzzyDateRegexField` class
by going into more detail about what it's capable
of, by referencing the default patterns that are used/
fallen back on when no custom patterns are provided
(incl. mention of the expected German date format), and
by explaining the existing example for a custom pattern
better (incl. giving a RegEx example to achieve it).

Resolves #66.